+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [In Development](https://github.com/ppfeufer/eve-online-killboard-widget/tree/development)
 [Full Changelog](https://github.com/ppfeufer/eve-online-killboard-widget/compare/v0.22.5...development)
+### Added
+- Highlight for solo kills in zkillboard style
+
 ### Changed
 - Cache handling is now done with an own database table instead of WP transient cache. This should keep the wp_options table clean.