+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 # Change Log
 
 ## [In Development](https://github.com/ppfeufer/eve-online-killboard-widget/tree/development)
-[Full Changelog](https://github.com/ppfeufer/eve-online-killboard-widget/compare/v0.22.5...development)
+[Full Changelog](https://github.com/ppfeufer/eve-online-killboard-widget/compare/v0.23.0...development)
+- in development
+
+## [0.23.0](https://github.com/ppfeufer/eve-online-killboard-widget/releases/tag/v0.23.0) - 2018-09-14
+[Full Changelog](https://github.com/ppfeufer/eve-online-killboard-widget/compare/v0.22.5...v0.23.0)
 ### Added
 - Highlight for solo kills in zkillboard style